+++ hardhat.config.ts
@@ -7,20 +7,19 @@ import 'hardhat-watcher'
 import '@openzeppelin/hardhat-upgrades'
 import '@openzeppelin/hardhat-defender'
 import '@nomiclabs/hardhat-etherscan'
-import { HardhatRuntimeEnvironment } from 'hardhat/types/runtime'
 
 import { HardhatUserConfig, task } from 'hardhat/config'
 
 import testUpgrade from './tasks/testUpgrade'
-
-
-
 
 require('dotenv').config()
 
 const DEPLOYER_PRIVATE_KEY = process.env.DEPLOYER_PRIVATE_KEY
 const ETHERSCAN_API_KEY = process.env.ETHERSCAN_API_KEY
 const OPTIMISTIC_ETHERSCAN_API_KEY = process.env.OPTIMISTIC_ETHERSCAN_API_KEY
+const OPTIMISM_GOERLI_ETHERSCAN_API_KEY =
+  process.env.OPTIMISM_GOERLI_ETHERSCAN_API_KEY
+const POLYGON_ETHERSCAN_API_KEY = process.env.POLYGON_ETHERSCAN_API_KEY
 
 function getRemappings() {
   return fs
@@ -31,7 +30,6 @@ function getRemappings() {
 }
 
 task('testUpgrade', 'Example task').setAction(testUpgrade)
-
 
 const config: HardhatUserConfig = {
   defender: {
@@ -52,28 +50,31 @@ const config: HardhatUserConfig = {
       url: `${process.env.GOERLI_RPC_URL}`,
       accounts: [`0x${DEPLOYER_PRIVATE_KEY}`],
       chainId: 5,
-      timeout: 20000,
     },
     optimisticEthereum: {
       url: `${process.env.OPTIMISM_RPC_URL}`,
       accounts: [`0x${DEPLOYER_PRIVATE_KEY}`],
       chainId: 10,
-      timeout: 20000,
     },
     mainnet: {
       url: `${process.env.MAINNET_RPC_URL}`,
       accounts: [`0x${DEPLOYER_PRIVATE_KEY}`],
       chainId: 1,
-      timeout: 20000,
-    },    
+    },
+    optimismGoerli: {
+      url: `${process.env.OPTIMISM_GOERLI_RPC_URL}`,
+      accounts: [`0x${DEPLOYER_PRIVATE_KEY}`],
+      chainId: 420,
+    },
+    polygon: {
+      url: `${process.env.POLYGON_RPC_URL}`,
+      accounts: [`0x${DEPLOYER_PRIVATE_KEY}`],
+      chainId: 137,
+    },
   },
   paths: {
     sources: './src', // Use ./src rather than ./contracts as Hardhat expects
     cache: './cache_hardhat', // Use a different cache for Hardhat than Foundry
-  },
-  // if a test takes longer than 40s Hardhat will fail it
-  mocha: {
-    timeout: 100000000,
   },
   watcher: {
     test: {
@@ -87,6 +88,8 @@ const config: HardhatUserConfig = {
       optimisticEthereum: OPTIMISTIC_ETHERSCAN_API_KEY!,
       goerli: ETHERSCAN_API_KEY!,
       mainnet: ETHERSCAN_API_KEY!,
+      optimismGoerli: OPTIMISM_GOERLI_ETHERSCAN_API_KEY!,
+      polygon: POLYGON_ETHERSCAN_API_KEY!,
     },
   },
   // This fully resolves paths for imports in the ./lib directory for Hardhat