+++ scripts/createProposal.js
@@ -19,6 +19,9 @@ const {
   MAINNET_BADGES_ADDRESS,
   MAINNET_RAFT_ADDRESS,
   MAINNET_SPECDATAHOLDER_ADDRESS,
+  SEPOLIA_BADGES_ADDRESS,
+  SEPOLIA_RAFT_ADDRESS,
+  SEPOLIA_SPECDATAHOLDER_ADDRESS,  
 } = process.env
 
 async function createProposal() {
@@ -51,6 +54,12 @@ async function createProposal() {
         } else if (network === 'mainnet') {
           contract.address = MAINNET_BADGES_ADDRESS
           contract.network = 'mainnet'
+        } else if (network === 'polygon') {
+          contract.address = POLYGON_BADGES_ADDRESS
+          contract.network = 'polygon'
+        } else if (network === 'sepolia') {
+          contract.address = SEPOLIA_BADGES_ADDRESS
+          contract.network = 'sepolia'
         }
         break
       case 'raft':
@@ -66,6 +75,12 @@ async function createProposal() {
         } else if (network === 'mainnet') {
           contract.address = MAINNET_RAFT_ADDRESS
           contract.network = 'mainnet'
+        } else if (network === 'polygon') {
+          contract.address = POLYGON_RAFT_ADDRESS
+          contract.network = 'polygon'
+        } else if (network === 'sepolia') {
+          contract.address = SEPOLIA_RAFT_ADDRESS
+          contract.network = 'sepolia'
         }
         break
       case 'specDataHolder':
@@ -81,6 +96,12 @@ async function createProposal() {
         } else if (network === 'mainnet') {
           contract.address = MAINNET_SPECDATAHOLDER_ADDRESS
           contract.network = 'mainnet'
+        } else if (network === 'polygon') {
+          contract.address = POLYGON_SPECDATAHOLDER_ADDRESS
+          contract.network = 'polygon'
+        } else if (network === 'sepolia') {
+          contract.address = SEPOLIA_SPECDATAHOLDER_ADDRESS
+          contract.network = 'sepolia'
         }
         break
       default:
@@ -95,9 +116,8 @@ async function createProposal() {
         viaType = 'Gnosis Safe'
         break
       case 'optimism-goerli':
-        // use env var
         // gnosis safe doesnt support optimism-goerli, so we need an address here
-        via = '0x76D84163bc0BbF58d6d3F2332f8A9c5B339dF983'
+        via = '0x21b02E9131D1BADE784a7874967DCb8Ef243F2A4'
         viaType = 'EOA'
         break
       case 'optimism':
@@ -108,6 +128,11 @@ async function createProposal() {
         via = MAINNET_GNOSIS_SAFE
         viaType = 'Gnosis Safe'
         break
+      case 'sepolia':
+        // gnosis safe doesnt support Sepolia, so we need an address here
+        via = '0x21b02E9131D1BADE784a7874967DCb8Ef243F2A4'
+        viaType = 'EOA'
+        break        
       default:
         throw new Error('Invalid network')
     }